requirements.txt
pandas
scikit-learn
numpy
streamlit
matplotlib
seaborn
scikit-image
Pillow
joblib
opencv-python